:construction_worker: Disable linting as part of react-scripts
See CRA upstream issue: https://github.com/facebook/create-react-app/issues/13271

Essentially, the eslint.config.mjs flat config is not picked up by
CRA, and it tries to run it with its default configuration (especially
now that .eslintrc.json is gone).

The linter errors during builds on the main branch are turned into
warnings (because of how the CRA shared config works) and currently
they're ignore anyway - I suspect that's why CI=false is added in the
frontend build in Github actions, as the default behaviour is to fail
the build.

Instead, the CI job now explicitly invokes 'npm run lint', which *will*
error the build. 'npm start' also disables the eslint plugin which is
a slightly worse developer experience, but the linting runs as a husky
pre-commit hook so hopefully this is an acceptable tradeoff.

ESLint flat config *is* the way forward and CRA is holding projects
back - if it becomes too problematic I'd consider migrating to ViteJS.