:hammer: [#582] Replace ESLint config with flat config
CRA ships with a default ESLint config, which is overridden in the
project to disable some bits. ESLint is moving towards flat config,
and CRA (unmaintained) is not compatible with it, so we set up our own
flows.

The envvar disables the CRA integration in react-scripts, as it does
not pick up the new config file. The pre-commit checks and CI pipeline
ensure that code is properly linted before being committed.

We also use the shared config from Maykin for consistency across
projects.