:sparkles: [#841] Added new endpoint to update the assignees
We are now updating a DestructionListAssignee of a list, and not necessarily the destruction_list.assignee. This meant that providing the role that the user should have was needed,
otherwise you don't know if you need to replace the author, a reviewer or the archivist. The reassing endpoint is deprecated.